django/requirements.txt
amqp==2.3.2
anyjson==0.3.3
appnope==0.1.0
argh==0.26.2
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
bcrypt==3.1.4
behave==1.2.6
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.2.2
decorator==4.3.0
defusedxml==0.5.0
dj-inmemorystorage==1.4.1
Django==2.1
django-allauth==0.36.0
django-appconf==1.0.2
django-celery-email==2.0.0
django-cors-headers==2.3.0
django-modeltranslation==0.13-beta1
django-redis==4.9.0
django-rest-auth==0.9.3
django-rosetta==0.8.1
django-simplefeedback==0.8
django-simple-history==2.4.0
django-ordered-model==2.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
Fabric3==1.14.post1
flake8==3.5.0
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
kombu==4.2.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
microsofttranslator==0.8
mock==2.0.0
more-itertools==4.2.0
oauthlib==2.1.0
paramiko==2.4.1
parse==1.8.4
parse-type==0.4.2
parso==0.3.0
pathtools==0.1.2
pbr==4.1.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.6.0
polib==1.1.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.5.4
pyasn1==0.4.3
pycodestyle==2.3.1
pycountry==18.5.26
pycparser==2.18
pycrypto==2.6.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.9.2
PyNaCl==1.2.1
pytest==3.6.3
pytest-cov==2.5.1
pytest-django==3.3.2
pytest-testmon==0.9.12
pytest-watch==4.2.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
raven==6.9.0
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
watchdog==0.8.3
wcwidth==0.1.7
wrapt==1.10.11
